% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableAttributes}
\alias{availableAttributes}
\title{List Attributes that can be Aggregated}
\usage{
availableAttributes(log, check = FALSE)
}
\arguments{
\item{log}{\verb{[ecr_logger]} ecr log object.}

\item{check}{\verb{[logical(1)]} whether to check consistency.}
}
\value{
\verb{[character]} attributes of individuals in population.
}
\description{
Looks at the first individuum in the first generation and
returns its attributes. If \code{check} is TRUE it checks that the
log object is consistent and throws an error if not. "Consistent"
here means that all individuals in all generations have the same
attributes.
}
