% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectorcpo.R
\name{cpoSelector}
\alias{cpoSelector}
\title{CPO that Selects Features}
\usage{
cpoSelector(selection, id, export = "export.default",
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
CPO that Selects Features
}
