% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{create.hypersphere.data}
\alias{create.hypersphere.data}
\title{Create Hypersphere Data}
\usage{
create.hypersphere.data(dim, n, dist = function(x) runif(x, -1, 1),
  norm = 2, radius = 1)
}
\arguments{
\item{dim}{\code{[integer(1)]} number of columns to create}

\item{n}{\code{[integer(1)]} number of sample to create}

\item{dist}{\code{[function]} function \code{n} -> \code{numeric(n)} that is used to sample points dimension-wise}

\item{norm}{\code{[numeric(1)]} Norm Exponent}

\item{radius}{\code{[numeric(1)]} Radious to check against}
}
\value{
\code{list(X = [Matrix], Y = [vector], orig.features = logical)}
}
\description{
Creates matrix X and vector Y:
\itemize{
\item Sample \code{dim} columns of data from \code{dist} (which must be a function
\code{n -> vector length (n)} and should probably sample randomly to create \code{X}.
\item \code{Y[i]} is \code{+1} if the L_\code{norm}-norm of \code{X[i, ]} is < \code{radius}, and \code{-1} otherwise.
}
}
\seealso{
Other Artificial Datasets: \code{\link{clonetask}},
  \code{\link{create.linear.data}},
  \code{\link{create.linear.toy.data}},
  \code{\link{create.regr.task}},
  \code{\link{task.add.permuted.cols}},
  \code{\link{task.add.random.cols}}
}
\concept{Artificial Datasets}
