% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fitnesses}
\alias{fitnesses}
\title{Extract Fitnesses from ECR Log}
\usage{
fitnesses(results, trafo = identity)
}
\arguments{
\item{results}{\verb{[ecr_multi_objective_result]} ecr run log.}

\item{trafo}{\verb{[function]} function \code{matrix}|\code{data.frame} -> \code{matrix}|\code{data.frame}
to transforms individual generation matrices.}
}
\value{
\code{data.frame} of fitnesses from ecr run log, with extra column \code{iter}.
}
\description{
Extract fitnesses for each generation from ECR log.
}
\seealso{
Other Utility Functions: 
\code{\link{paretoEdges}()}
}
\concept{Utility Functions}
