% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPopulations}
\alias{getPopulations}
\title{Get Populations}
\usage{
getPopulations(log)
}
\arguments{
\item{log}{\verb{[ecr_logger]} ecr log object}
}
\value{
\code{list} of populations.
}
\description{
Get populations from ecr_logger. Replaces ecr::getPopulation because
original is buggy.
}
