% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStatistics}
\alias{getStatistics}
\title{Get Statistics}
\usage{
getStatistics(log)
}
\arguments{
\item{log}{\code{[ecr_logger]} ecr log object}
}
\description{
Get statistics from ecr_logger. Replaces ecr::getStatistics
because original is buggy.
}
