% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initSelector}
\alias{initSelector}
\title{Initialize Selector}
\usage{
initSelector(individuals, distribution = function() floor(runif(1, 0,
  length(individuals[[1]]$selector.selection) + 1)), soften.op = NULL,
  soften.op.strategy = NULL, soften.op.repeat = 1,
  reject.zero = TRUE)
}
\arguments{
\item{individuals}{\code{[list of named lists]} the individuals to initialize}

\item{distribution}{\code{[function]} function that returns a random integer
from 0 to the length of each individual's \code{$selector.selection} slot.
Defaults to the uniform distribution from 1 to \code{length()}.}

\item{soften.op}{\code{[ecr_mutator]} an optional mutator to apply to the
\code{$selector.selection} variable.}

\item{soften.op.strategy}{\code{function} an optional function that can set
the \code{soften.op}'s parameters. See \code{\link{combine.operators}} strategy parameters.
Ignored if \code{soften.op} is not given.}

\item{soften.op.repeat}{\code{[integer(1)]} how often to repeat \code{soften.op}
application. Ignored if \code{soften.op} is not given.}

\item{reject.zero}{\code{[logical(1)]} whether to reject (and sample anew)
vectors with all zeroes. Default \code{TRUE}.}
}
\value{
\code{list of named lists} The individuals with initialized
`$selector.selection.
}
\description{
Sample the \code{selector.selection} variable such that the number of ones
has a given distribution.
}
