% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{intifyMutator}
\alias{intifyMutator}
\alias{intifyRecombinator}
\title{Turn Continuous-Space Operators into Integer-Space Operators}
\usage{
intifyMutator(operator)

intifyRecombinator(operator)
}
\arguments{
\item{operator}{\code{[ecr_operator]} \code{\link[ecr:makeOperator]{ecr_operator}} that supports continuous
variables.}
}
\value{
\code{\link[ecr:makeOperator]{ecr_operator}} operator that operates on integers.
}
\description{
The input \code{operator} is wrapped: individuals are fed to it as-is, and
output is rounded. Upper and lower bounds are both shifted by 0.5 down or
up, respectively, to retain a fair distribution.
}
