% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtermat.R
\name{makeFilterMat}
\alias{makeFilterMat}
\title{Create a Filter-Matrix}
\usage{
makeFilterMat(task, filters, expectfeats = getTaskNFeats(task)/2,
  minprob = 0, maxprob = 1)
}
\arguments{
\item{task}{\code{[Task]} The task to generate filter information for}

\item{filters}{\code{[character]} The filters to use}

\item{expectfeats}{\code{[integer(1)]} The expected number of features to have in equilibrium.}

\item{minprob}{\code{[numeric(1)]} The minimum probability for each feature}

\item{maxprob}{\code{[numeric(1)]} The maximum probability for each feature}
}
\value{
\code{matrix}
}
\description{
A Filter-Matrix can be used in combination with \code{\link{mutUniformMetaReset}} for
heuristic-supported biased mutation.
}
