% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{makeFilterStrategy}
\alias{makeFilterStrategy}
\title{Create a Filter Strategy Function}
\usage{
makeFilterStrategy(reset.dists, weight.param.name)
}
\arguments{
\item{reset.dists}{\verb{[matrix]} see \code{reset.dists} in \link{mutUniformMetaReset}.}

\item{weight.param.name}{\verb{[character(1)]} name of parameter to use as
\code{reset.dist.weights} in \link{mutUniformMetaReset}.}
}
\value{
\code{function}
}
\description{
Creates a strategy function that uses the \code{weight.param.name} entry of
individuals as a weighting vector \code{reset.dist.weights} and \code{reset.dists}
for \code{\link{mutUniformMetaReset}} and \code{\link{mutUniformMetaResetSHW}}.
}
