% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customnsga2.R
\name{mosmafs.nsga2}
\alias{mosmafs.nsga2}
\title{NSGA2 without Prejudice}
\usage{
mosmafs.nsga2(...)
}
\arguments{
\item{...}{Same parameters as in \link[ecr:ecr]{ecr::ecr}}
}
\description{
Performs NSGA2 optimization as done by the ecr2-package, but with
arguments that may be overwritten.
}
