% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutGaussInt}
\alias{mutGaussInt}
\title{Integer Gaussian Mutator}
\usage{
mutGaussInt(ind, ..., lower, upper)
}
\description{
See \link[ecr:mutGauss]{ecr::mutGauss}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recIntSBX}}
}
\concept{operators}
