% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutGaussIntScaled}
\alias{mutGaussIntScaled}
\title{Integer Scaled Gaussian Mutator}
\usage{
mutGaussIntScaled(ind, ..., lower, upper)
}
\description{
See \link{mutGaussScaled}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recIntSBX}}
}
\concept{operators}
