% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutGaussScaled}
\alias{mutGaussScaled}
\title{Integer Scaled Gaussian Mutator}
\usage{
mutGaussScaled(ind, p = 1, sdev = 0.05, lower, upper)
}
\description{
See \link[ecr:mutGauss]{ecr::mutGauss}. Allows a vector of standard deviations. Scales
standard deviations to the range of \code{[lower, upper]}.
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recIntSBX}}
}
\concept{operators}
