% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutPolynomialInt}
\alias{mutPolynomialInt}
\title{Integer Polynomial Mutator}
\usage{
mutPolynomialInt(ind, ..., lower, upper)
}
\description{
See \link[ecr:mutPolynomial]{ecr::mutPolynomial}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recGaussian}},
  \code{\link{recIntIntermediate}}, \code{\link{recIntSBX}}
}
\concept{operators}
