% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutUniformInt}
\alias{mutUniformInt}
\title{Integer Uniform Mutator}
\usage{
mutUniformInt(ind, ..., lower, upper)
}
\description{
See \link[ecr:mutUniform]{ecr::mutUniform}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}}, \code{\link{recGaussian}},
  \code{\link{recIntIntermediate}}, \code{\link{recIntSBX}}
}
\concept{operators}
