% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutUniformMetaReset}
\alias{mutUniformMetaReset}
\title{Parametrised Uniform Reset for Binary Parameters}
\usage{
mutUniformMetaReset(ind, p = 0.1, reset.dists, reset.dist.weights)
}
\arguments{
\item{ind}{\code{[integer]} binary individuum with values 0 or 1}

\item{p}{\code{[numeric(1)]} entry-wise reset probability}

\item{reset.dists}{\code{[matrix]} columns of probabilities to draw 1-bit per entry, if reset is performed.
Must have \code{length(ind)} columns and \code{length(reset.dist.weights)} rows.}

\item{reset.dist.weights}{\code{[numeric]} weight vector to select among \code{reset.dists} columns.}
}
\value{
\code{[integer]} the mutated individuum.
}
\description{
perform \code{\link{mutUniformReset}}, with \code{reset.dist = reset.dists \%*\% reset.dist.weights}.
}
