% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutUniformParametric}
\alias{mutUniformParametric}
\alias{mutUniformParametricScaled}
\alias{mutUniformParametricInt}
\alias{mutUniformParametricIntScaled}
\title{Parametric Uniform  Mutation}
\usage{
mutUniformParametric(ind, p, lx, lower, upper)

mutUniformParametricScaled(ind, p, sdev, lower, upper)

mutUniformParametricInt(ind, ..., lower, upper)

mutUniformParametricIntScaled(ind, ..., lower, upper)
}
\arguments{
\item{ind}{\code{[numeric | integer]} individuum}

\item{p}{\code{[numeric]} probability to affect each individual component}

\item{lower}{\code{[numeric]} lower bound}

\item{upper}{\code{[numeric]} upper bound}

\item{`lx`}{\code{[numeric]} uniform distribution bandwidth}

\item{`sdev`}{\code{[numeric]} standard deviation, will be scaled to \code{upper - lower}}
}
\description{
Adds a variable \code{delta} to each component \code{ind[i]}
with probability \code{p}, where \code{delta} is uniformly
distributed between \code{pmax(lower - ind[x], -lx/2)} and
\code{pmin(upper - ind[i], lx/2)}.
}
