% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{paretoEdges}
\alias{paretoEdges}
\title{Get Pareto Front Edges from Fitness Matrix}
\usage{
paretoEdges(fitness, refpoint)
}
\arguments{
\item{fitness}{\verb{[matrix | data.frame]} matrix or (numeric) \code{data.frame}
with two columns and rows for each individuum.}

\item{refpoint}{\verb{[numeric(2)]} reference point.}
}
\value{
\code{data.frame} with three columns: The points on the pareto front,
and a \code{logical} column \code{point} indicating whether the point is on the pareto front
(\code{TRUE}) or an auxiliary point for plotting (\code{FALSE}).
}
\description{
Get the edges defining a 2D pareto front for plotting.
}
\seealso{
Other Utility Functions: 
\code{\link{fitnesses}()}
}
\concept{Utility Functions}
