% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{recIntSBX}
\alias{recIntSBX}
\title{Integer SBX Recombinator}
\usage{
recIntSBX(inds, ..., lower, upper)
}
\description{
See \link[ecr:recSBX]{ecr::recSBX}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}
}
\concept{operators}
