% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{selSimpleUnique}
\alias{selSimpleUnique}
\title{Simple Selector without Replacement}
\usage{
selSimpleUnique(fitness, n.select)
}
\arguments{
\item{fitness}{\code{[matrix]} fitness matrix, one column per individuum}

\item{n.select}{\code{[integer(1)]} number of individuums to select}
}
\description{
Simple Selector without Replacement
}
\seealso{
Other Selectors: \code{\link{selTournamentMO}}
}
\concept{Selectors}
