% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setMosmafsVectorized}
\alias{setMosmafsVectorized}
\title{Set or change attribute \code{mosmafs.vectorize} in fitness function}
\usage{
setMosmafsVectorized(fn, vectorize = TRUE)
}
\description{
Set or change attribute \code{mosmafs.vectorize} in fitness function
}
