% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customnsga2.R
\name{slickEcr}
\alias{slickEcr}
\title{Modified Interface to ECR}
\usage{
slickEcr(fitness.fun, lambda, population, mutator, recombinator,
  generations = 100, parent.selector = selSimple,
  survival.selector = selNondom, p.recomb = 0.7, p.mut = 0.3,
  survival.strategy = c("plus", "comma"), n.elite = 0,
  fidelity = NULL, log.stats = list(fitness = list("min", "mean",
  "max")))
}
\arguments{
\item{fitness.fun}{\code{[smoof_multi_objective_function]} fitness function, must be a
"\code{smoof}" function.}

\item{lambda}{\code{[integer(1)]} number of individuals to add each generation}

\item{population}{\code{[list]} list of individuals to start off from.}

\item{mutator}{\code{[ecr_mutator]} mutation operator}

\item{recombinator}{\code{[ecr_recombinator]} recombination operator}

\item{generations}{\code{[integer(1)]} number of iterations to evaluate}

\item{parent.selector}{\code{[ecr_selector]} parent selection operator}

\item{survival.selector}{\code{[ecr_selector]} survival selection operator}

\item{p.recomb}{\code{[numeric(1)]} probability to apply a recombination operator}

\item{p.mut}{\code{[numeric(1)]} probability to apply mutation operator}

\item{survival.strategy}{\code{[character(1)]} one of \code{"plus"} or \code{"comma"}}

\item{n.elite}{\code{[integer(1)]} Number of elites to keep, only used if
\code{survival.strategy} is \code{"comma"}}

\item{fidelity}{\code{[data.frame | NULL]} If this is given, it controls the
fidelity of the function being evaluated, via its \code{fidelity} argument.
It must then be a \code{data.frame} with two or three columns. The first column
gives the generation at which the fidelity first applies; the second
column controls the fidelity at that generation or later; the third column,
if given, controls the additional fidelity whenever the result of the first
evaluation is not dominated by any result of the previous generation. The
entries in the first column must be strictly ascending. The first element
of the first column must always be 1. Whenever fidelity changes, the whole
population is re-evaluated, so it is recommended to use only few different
fidelity jumps throughout all generations.}

\item{log.status}{\code{[list]} information to log}
}
\description{
Mostly \code{\link[ecr:ecr]{ecr::ecr}}, with some simplifications and extensions.
}
