% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unbiasedHoldoutDomHV}
\alias{unbiasedHoldoutDomHV}
\title{Unbiased Dominated Hypervolume on Holdout Data}
\usage{
unbiasedHoldoutDomHV(fitness, holdout, refpoint)
}
\arguments{
\item{fitness}{\code{[matrix]} fitness matrix on training data}

\item{holdout}{\code{[matrix]} fitness matrix on holdout data}

\item{refpoint}{\code{numeric} reference point}
}
\value{
\code{numeric}
}
\description{
Calculate dominated hypervolume on holdout data. The result is
unbiased with respect to (uncorrelated w/r/t objectives) noise in holdout data
performance.

Only works on 2-objective performance matrices
}
