% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{valuesFromNames}
\alias{valuesFromNames}
\title{Convert Discrete Parameters from Names to Values}
\usage{
valuesFromNames(paramset, value)
}
\arguments{
\item{paramset}{\code{[ParamSet]} The \code{\link[ParamHelpers:ParamSet]{ParamHelpers::ParamSet}} used to generate the value.}

\item{value}{\code{[named list]} Names list of parameters sampled from \code{paramset}.}
}
\value{
\code{named list} of parameter values, with \code{character} entries representing names
of values of discrete params converted to the actual values.
}
\description{
Convert parameter values sampled with \code{\link[ParamHelpers:sampleValue]{ParamHelpers::sampleValue()}} and
\code{discrete.names = TRUE} to true parameter values.
}
