% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{create.linear.data}
\alias{create.linear.data}
\title{Create Linear Model Data}
\usage{
create.linear.data(n, p, q = exp(-1), beta0 = 1, rho = 0,
  permute = TRUE)
}
\arguments{
\item{n}{\code{[integer(1)]} number of rows to generate}

\item{p}{\code{[integer(1)]} number of columns to generate}

\item{q}{\code{[numeric(1)]} attenuation factor for beta coefficients}

\item{beta0}{\code{[numeric(1)]} size of first coefficient}

\item{rho}{\code{[numeric(1)]} parameter for correlation matrix}

\item{permute}{\code{[logical(1)]} whether to permute columns of \code{X} and coefficient
vector (\code{beta}).}
}
\value{
\code{list(X=[Matrix], Y=[vector], beta = [vector], orig.features = logical)}
}
\description{
Create linear model data \code{Y = X * beta + epsilon} with
\code{X = X'} if \code{[permute] == FALSE}, and \code{X = shuffle_columns(X')}
if \code{[permute] == TRUE} where \code{X'} is an \code{[n] * [p]} matrix of MVN distributed
rows with covariance matrix\preformatted{1     rho rho^2 rho^3 ... rho^p
rho     1 rho   rho^2 ... rho^(p-1)
rho^2 rho   1   rho ...   rho^(p-2)
...
rho^p ...
}

\code{epsilon} is normally distributed with stdev 1.

\code{beta = shuffle(beta')} if \code{[permute] == TRUE} and \code{beta = beta'} otherwise,
and \code{beta'[i] = beta0 * q ^ (i - 1)} for \code{i} = \code{1..p}

\code{$orig.features} are the features with \code{beta > 1 / sqrt(n)}
}
\seealso{
Other Artificial Datasets: \code{\link{clonetask}},
  \code{\link{create.hypersphere.data}},
  \code{\link{create.linear.toy.data}},
  \code{\link{create.regr.task}},
  \code{\link{task.add.permuted.cols}},
  \code{\link{task.add.random.cols}}
}
\concept{Artificial Datasets}
