% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{create.regr.task}
\alias{create.regr.task}
\alias{create.classif.task}
\title{Create mlr-Task from Data}
\usage{
create.regr.task(id, data)

create.classif.task(id, data, cutoff = 0)
}
\arguments{
\item{id}{\verb{[character(1)]} ID to use for \code{\link{Task}}.}

\item{data}{\verb{[named list]} with columns entries \verb{$X}, \verb{$Y}, and \verb{$orig.features}.}

\item{cutoff}{\verb{[numeric(1)]} cutoff at which to binarize target.}
}
\value{
\code{\link{Task}}
}
\description{
Both \code{create.regr.task} and \code{create.classif.task} take a numeric target column
\code{Y}, but \code{create.classif.task} binarizes it on \code{cutoff} to create a classification
task, while \code{create.regr.task} creates a regression task.
}
\seealso{
Other Artificial Datasets: 
\code{\link{clonetask}()},
\code{\link{create.hypersphere.data}()},
\code{\link{create.linear.data}()},
\code{\link{create.linear.toy.data}()},
\code{\link{task.add.permuted.cols}()},
\code{\link{task.add.random.cols}()}

Other Artificial Datasets: 
\code{\link{clonetask}()},
\code{\link{create.hypersphere.data}()},
\code{\link{create.linear.data}()},
\code{\link{create.linear.toy.data}()},
\code{\link{task.add.permuted.cols}()},
\code{\link{task.add.random.cols}()}
}
\concept{Artificial Datasets}
