% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStatistics}
\alias{getStatistics}
\title{Get Statistics}
\usage{
getStatistics(log)
}
\arguments{
\item{log}{\verb{[ecr_logger]} ecr log object}
}
\value{
\code{data.frame} of logged statistics.
}
\description{
Get statistics from ecr_logger. Replaces \code{\link[ecr:getStatistics]{ecr::getStatistics}}
because original is buggy.
}
