% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminators.R
\name{mosmafsTermEvals}
\alias{mosmafsTermEvals}
\alias{mosmafsTermTime}
\alias{mosmafsTermFidelity}
\alias{mosmafsTermStagnationHV}
\alias{mosmafsTermStagnationObjMean}
\title{Termination Function Creator}
\usage{
mosmafsTermEvals(evals)

mosmafsTermTime(time)

mosmafsTermFidelity(fidelity)

mosmafsTermStagnationHV(staggen)

mosmafsTermStagnationObjMean(staggen)
}
\arguments{
\item{evals}{\code{[integer(1)]} limit evals}

\item{time}{\code{[numeric(1)]} limit evaluation time (which does not count
holdout fitting time}

\item{fidelity}{\code{[numeric(1)]} total fidelity evaluation to limit}

\item{generations}{\code{[integer(1)]} limit generations. Initial population does
not count.}

\item{staggens}{\code{[integer(1)]} number of generations without progress in
hypervolume or mean objective value}
}
\value{
\code{function} a terminator function
}
\description{
These create functions that can be given to \code{slickEcr}'s \code{generations} argument

The stagnation terminators only count stagnation from the last time
the fidelity was changed in a way that led to population re-evaluation.
}
