% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutDoubleGeom}
\alias{mutDoubleGeom}
\title{Double Geometric Distribution Mutator}
\usage{
mutDoubleGeom(ind, p = 1, geomp = 0.9, lower, upper)
}
\arguments{
\item{ind}{\code{[integer]} individuum to mutate}

\item{p}{\code{[numeric(1)]} probability with which to mutate}

\item{geomp}{\code{[numeric(1)]} geometric distribution parameter}

\item{lower}{\code{[integer]} lower bounds on \code{ind} values. May have same length as
\code{ind} or may be shorter, in which case it is recycled.}

\item{upper}{\code{[integer]} upper bounds on \code{ind} values. May have same length as
\code{ind} or may be shorter, in which case it is recycled.}
}
\value{
\code{\link{integer}}
}
\description{
"Double Geometric" mutation operator for integer parameters: with prob. \code{p} both
adds a random geometrically distributed value, and subtracts (a different) one.
}
\seealso{
Other operators: \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recIntSBX}}
}
\concept{operators}
