% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutGaussIntScaled}
\alias{mutGaussIntScaled}
\title{Integer Scaled Gaussian Mutator}
\usage{
mutGaussIntScaled(ind, ..., lower, upper)
}
\arguments{
\item{ind}{\code{[integer]} integer vector/individual to mutate.}

\item{...}{other arguments passed on to the method.}

\item{lower}{\code{[integer]} vector of minimal values for each parameter of the decision space. Must have the same length as \code{ind}.}

\item{upper}{\code{[integer]} vector of maximal values for each parameter of the decision space. Must have the same length as \code{ind}.}
}
\description{
See \link{mutGaussScaled}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutPolynomialInt}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recGaussian}},
  \code{\link{recIntIntermediate}}, \code{\link{recIntSBX}}
}
\concept{operators}
