% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{mutPolynomialInt}
\alias{mutPolynomialInt}
\title{Integer Polynomial Mutator}
\usage{
mutPolynomialInt(ind, ..., lower, upper)
}
\arguments{
\item{ind}{\code{[integer]} integer vector/individual to mutate.}

\item{lower}{\code{[integer]} vector of minimal values for each parameter of the
decision space. Must have the same length as \code{ind}.}

\item{uppper}{\code{[integer]} vector of maximal values for each parameter of the
decision space. Must have the same length as \code{ind}.}
}
\description{
See \link[ecr:mutPolynomial]{ecr::mutPolynomial}
}
\seealso{
Other operators: \code{\link{mutDoubleGeom}},
  \code{\link{mutGaussIntScaled}},
  \code{\link{mutGaussInt}}, \code{\link{mutGaussScaled}},
  \code{\link{mutRandomChoice}},
  \code{\link{mutUniformInt}}, \code{\link{recGaussian}},
  \code{\link{recIntIntermediate}}, \code{\link{recIntSBX}}
}
\concept{operators}
