% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naiveHoldoutDomHV}
\alias{naiveHoldoutDomHV}
\title{Naive Hypervolume on Holdout Data}
\usage{
naiveHoldoutDomHV(fitness, holdout, refpoint)
}
\arguments{
\item{fitness}{\verb{[matrix]} fitness matrix of training data.}

\item{holdout}{\verb{[matrix]} fitness matrix of holdout data.}

\item{refpoint}{\verb{[numeric]} reference point.}
}
\value{
\code{numeric}
}
\description{
Calculate dominated hypervolume on holdout data. The result is
biased depending on noise in holdout data performance.
}
