% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{selSimpleUnique}
\alias{selSimpleUnique}
\title{Simple Selector without Replacement}
\usage{
selSimpleUnique(fitness, n.select)
}
\arguments{
\item{fitness}{\verb{[matrix]} fitness matrix, one column per individual.}

\item{n.select}{\verb{[integer(1)]} number of individuals to select.}
}
\value{
\verb{[matrix]} selected individuals.
}
\description{
Simple Selector without Replacement
}
\seealso{
Other Selectors: 
\code{\link{selTournamentMO}()}
}
\concept{Selectors}
