% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unbiasedHoldoutDomHV}
\alias{unbiasedHoldoutDomHV}
\title{Unbiased Dominated Hypervolume on Holdout Data}
\usage{
unbiasedHoldoutDomHV(fitness, holdout, refpoint)
}
\arguments{
\item{fitness}{\verb{[matrix]} fitness matrix of training data.}

\item{holdout}{\verb{[matrix]} fitness matrix of holdout data.}

\item{refpoint}{\verb{[numeric]} reference point.}
}
\value{
\code{numeric}
}
\description{
Calculate dominated hypervolume on holdout data. The result is
unbiased with respect to (uncorrelated with respect to objectives) noise in holdout data
performance, but it is \emph{not} an estimate of real "dominated hypervolume".

Only works on two-objective performance matrices.
}
